C#: Update to net7.0
From 4c1d30fa00cc9f4cdea9f2c90b55fd42ee498ae5

Because of ref safety changes in the languages, all methods that return an
interop struct have to have all other reference parameters marked as scoped
to signal that the method does not capture that reference.

The variant change is necessary, because for some reason a type of the exact
shape godot_variant is in, crashes the .NET 7 JIT, but when changing it to be
sequential with the same effective layout it works.